# api/src/schemas/user.js
const UserSchema = new dynamoose.Schema(
  {
    user_id: {
      type: String,
      hashKey: true,
      required: true,
    },
    wallets: {
      type: Array,
      schema: [{
        type: Object,
        schema: {
          wallet_address: String,
          wallet_public_key: String,
        }
      }],
      required: false,
    },
    nonce: {
      type: String,
      required: false,
      index: {
        name: 'nonceIndex',
        global: true,
      },
    },
    nickname: {
      type: String,
      required: false,
    },
    email: {
      type: String,
      required: false,
      index: {
        name: 'emailIndex',
        global: true,
      },
    },
    team_ids: {
      type: Array,
      schema: [String],
      required: false,
    },
    created_at: {
      type: Number,
      required: false,
    },
    created_by: {
      type: String,
      required: false,
    },
    updated_at: {
      type: Number,
      required: false,
    },
    updated_by: {
      type: String,
      required: false,
    },
    delete_flg: {
      type: Boolean,
      required: false,
    },
  },
  {
    timestamps: false,
  }
);

module.exports = UserSchema;