# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: timedex REST API definition
  description: REST API
  version: 1.0.0
servers:
  - url: http://localhost:50001
    description: API Mock
paths:
  /v1/auth/generate-nonce:
    post:
      security:
        - Bearer: []
      tags:
        - Authentication
      summary: Generate nonce
      operationId: generateNonce
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthNonceRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthNonceResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '500':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
...
  /v1/user/{user_id}:
    get:
      security:
        - Bearer: []
      tags:
        - User
      summary: Get user info
      operationId: getUser
      parameters:
        - name: Authorization
          in: header
          description: JWT
          required: true
          schema:
            type: string
          example: Bearer xxx
        - name: user_id
          in: path
          description: User ID
          required: true
          schema:
            type: string
          example: 9490ebf5-e378-4074-97d0-3ce5004ae918
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseNotFound'
        '500':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
    put:
      security:
        - Bearer: []
      tags:
        - User
      summary: Update new user
      operationId: updateUser
      parameters:
        - name: Authorization
          in: header
          description: JWT
          required: true
          schema:
            type: string
          example: Bearer xxx
        - name: user_id
          in: path
          description: User ID
          required: true
          schema:
            type: string
          example: 9490ebf5-e378-4074-97d0-3ce5004ae918
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              nickname: Yamata tarou
              email: yamata@web3-team.com
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
              example:
                message: Success
                id_info:
                  key: user_id
                  value: 9490ebf5-e378-4074-97d0-3ce5004ae918
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseForbidden'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseNotFound'
        '500':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
...
  /v1/send/transaction:
    post:
      security:
        - Bearer: []
      tags:
        - Transaction
      summary: Execute transaction
      operationId: sendTransaction
      parameters:
        - name: Authorization
          in: header
          description: JWT
          required: true
          schema:
            type: string
          example: Bearer xxx
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTransactionRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendTransactionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Credentials or access token for API
  schemas:
    AuthNonceRequest:
      type: object
      properties:
        wallet_address:
          type: string
          example: '0x328973ce1433d7593191271f75a428af55c6528f'
    AuthNonceResponse:
      type: object
      properties:
        nonce:
          type: string
          example: abcdef
        created_at:
          type: number
          example: 1677498728065
...
    CommonResponse:
      type: object
      properties:
        message:
          type: string
          example: Success
        id_info:
          type: object
          $ref: '#/components/schemas/KeyValue'
...
    KeyValue:
      type: object
      properties:
        key:
          type: string
          example: user_id
        value:
          type: string
          example: 2490ebf5-e378-4074-97d0-3ce5004ae9cg
...
    ResponseBadRequest:
      type: object
      properties:
        message:
          type: string
          example: Bad Request
...
    ResponseForbidden:
      type: object
      properties:
        message:
          type: string
          example: Forbidden
    ResponseInternalServerError:
      type: object
      properties:
        message:
          type: string
          example: Service Unavailable
    ResponseNotFound:
      type: object
      properties:
        message:
          type: string
          example: Not Found
    ResponseUnauthorized:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
    SendTransactionRequest:
      type: object
      properties:
        raw_transaction:
          type: string
          example: 0x001122...
        type:
          type: string
          example: Publish/CreateTeam/...
    SendTransactionResponse:
      type: object
      properties:
        message:
          type: string
          example: Success
        token_address:
          type: string
          example: 0x
        token_amount:
          type: string
          example: '10.15'
...
    User:
      type: object
      properties:
        user_id:
          type: string
          example: 9490ebf5-e378-4074-97d0-3ce5004ae918
        nickname:
          type: string
          example: Yamata tarou
        email:
          type: string
          example: user@web3-team.com
        teams_ids:
          type: array
          items:
            type: string
          example:
            - 9490ebf5-e378-4074-97d0-3ce5004ae918
            - 2490ebf5-e378-4074-97d0-3ce5004ae919
        created_at:
          type: number
          example: 1655562492284
        created_by:
          type: string
          example: 2490ebf5-e378-4074-97d0-3ce5004ae9cf
        updated_at:
          type: number
          example: 1655562492284
        updated_by:
          type: string
          example: 2490ebf5-e378-4074-97d0-3ce5004ae9cf
        delete_flg:
          type: boolean
          example: false
...